refactor(executor): Rename mysql executor to sql executor
