+++ eventbus/mysql/eventbus_test.go
@@ -28,7 +28,7 @@ import (
 	"gorm.io/gorm"
 
 	"github.com/bytedance/dddfirework"
-	exec_mysql "github.com/bytedance/dddfirework/executor/mysql"
+	exec_mysql "github.com/bytedance/dddfirework/executor/sql"
 	"github.com/bytedance/dddfirework/testsuit"
 )
